requirements.txt
aiohttp==3.9.1
beautifulsoup4==4.12.2
rich==13.7.0
lxml==4.9.3
supabase==2.3.4
python-dotenv==1.0.0
scikit-learn==1.3.2
numpy==1.24.3
openai==1.12.0
```

```

